Remove credentials data
Issue-ID: SDNC-458

Service Decomposition and Network Discovery Microservice have sensitive
keystore/certificate/password data in their configuration files that
needs to be removed.

In order for Network Discovery Microservice JUNIT to execute a dummy
Enricher REST client client-cert-onap.p12 was added to
src/test/resources. This is the ONAP certificate that will be used when
the ONAP Enricher is available, then it will be copied to config/auth.

Change-Id: I3aac665815757f9a31bbe0b4020f44cd3fe1eb12
Signed-off-by: Tait,Trevor(rt0435) <rtait@amdocs.com>