Update patch set 1
Patch Set 1:

This is failing due to a change in job configuration - it is now doing a docker build as part of verification, which means it needs to run on one of the docker build nodes (otherwise docker isn't installed, so build fails).  Submitted a review to ci-management to fix configuration.

Patch-set: 1